+++ record-and-playback/systemd/bbb-rap-sanity-worker.service
@@ -1,0 +1,10 @@
+[Unit]
+Description=BigBlueButton recording and playback sanity check worker
+ConditionPathExistsGlob=/var/bigbluebutton/recording/status/archived/*.done
+
+[Service]
+Type=simple
+ExecStart=/usr/local/bigbluebutton/core/scripts/rap-sanity-worker.rb
+WorkingDirectory=/usr/local/bigbluebutton/core/scripts
+User=tomcat7
+Slice=bbb_record_core.slice
