+++ record-and-playback/systemd/bbb_record_core.slice
@@ -1,0 +1,2 @@
+[Unit]
+Description = BigBlueButton Recording Processing
